fixed typo in icooutput.cpp and tgaoutput.cpp where they were set to change the format to UINT8 if they weren't UINT16, instead of UINT8

+++ src/ico.imageio/icooutput.cpp
@@ -198,7 +198,7 @@ ICOOutput::open (const std::string &name, const ImageSpec &userspec,
                     + (4 - ((m_spec.width + 7) / 8) % 4) % 4; // padding
 
         // Force 8 bit integers
-        if (m_spec.format != TypeDesc::UINT16)
+        if (m_spec.format != TypeDesc::UINT8)
             m_spec.set_format (TypeDesc::UINT8);
     }
